+++ server/component/style/dataConfigBuilder/dataConfigBuilderView.php
@@ -9,7 +9,7 @@ require_once __DIR__ . "/../formField/FormFieldView.php";
 /**
  * The view class of the ConditionBuilder style component.
  */
-class ConditionBuilderView extends FormFieldView
+class DataConfigBuilderView extends FormFieldView
 {
     /* Private Properties *****************************************************/
 
@@ -66,27 +66,33 @@ class ConditionBuilderView extends FormFieldView
     public function output_builder()
     {
         $modal = new BaseStyleComponent('modal', array(
-            'title' => "Condition Builder",
-            "css" => "condition_builder_modal_holder",
+            'title' => "Data Config Builder",
+            'id' => "data_config_builder_modal_holder",
+            "css" => "data_config_builder_modal_holder",
             'children' => array(
                 new BaseStyleComponent("div", array(
-                    "css" => "condition_builder"
-                )),
-                new BaseStyleComponent("div", array(
-                    "css" => "modal-footer",
+                    "css" => "d-flex flex-column justify-content-between data_config_builder_holder",
                     "children" => array(
-                        new BaseStyleComponent("button", array(
-                            "label" => "Save",
-                            "url" => "#",
-                            "type" => "secondary",
-                            "css" => "saveConditionBuilder bnt-sm"
+                        new BaseStyleComponent("div", array(
+                            "css" => "data_config_builder"
                         )),
-                    )
-                ))
-            ),
+                        new BaseStyleComponent("div", array(
+                            "css" => "modal-footer",
+                            "children" => array(
+                                new BaseStyleComponent("button", array(
+                                    "label" => "Save",
+                                    "url" => "#",
+                                    "type" => "secondary",
+                                    "css" => "saveDataConfig btn-sm"
+                                )),
+                            )
+                        ))
+                    ),
+                )),
+            )
         ));
         $modal->output_content();
-        require __DIR__ . "/tpl_condition_builder.php";
+        require __DIR__ . "/tpl_dataConfig_builder.php";
     }
 }
 ?>