+++ server/db/update_scripts/02_update_v4.9.0_v5.0.0.sql
@@ -317,4 +317,65 @@ WHERE `name` = 'condition';
 --  hide all fields `jquery_builder_json`
 UPDATE styles_fields
 SET hidden = 1
-WHERE id_fields = get_field_id('jquery_builder_json')
+WHERE id_fields = get_field_id('jquery_builder_json');
+
+-- Add new field type `data-config` 
+INSERT INTO `fieldType` (`id`, `name`, `position`) VALUES (NULL, 'data-config', '7');
+
+-- Add new style `dataConfigBuilder`
+INSERT INTO `styles` (`name`, `id_type`, id_group, description) VALUES ('dataConfigBuilder', '1', (select id from styleGroup where `name` = 'intern' limit 1), 'Internal style used for the `data-config` field');
+
+-- make field `data_config` from type `data-config`
+UPDATE `fields`
+SET id_type = (SELECT id FROM fieldType WHERE `name` = 'data-config' LIMIT 0,1)
+WHERE `name` = 'data_config';
+
+DELIMITER //
+
+DROP PROCEDURE IF EXISTS get_uploadTable_with_filter //
+
+CREATE PROCEDURE get_uploadTable_with_filter( table_id_param INT, filter_param VARCHAR(1000) )
+BEGIN
+    SET @@group_concat_max_len = 32000;
+    SET @sql = NULL;
+    SELECT
+    GROUP_CONCAT(DISTINCT
+        CONCAT(
+            'max(case when col.name = "',
+                col.name,
+                '" then value end) as `',
+            replace(col.name, ' ', ''), '`'
+        )
+    ) INTO @sql
+    FROM  uploadTables t
+	INNER JOIN uploadRows r on (t.id = r.id_uploadTables)
+	INNER JOIN uploadCells cell on (cell.id_uploadRows = r.id)
+	INNER JOIN uploadCols col on (col.id = cell.id_uploadCols)
+    WHERE t.id = table_id_param;
+
+    IF (@sql is null) THEN
+        SELECT table_name from view_uploadTables where 1=2;
+    ELSE
+        BEGIN
+            SET @sql = CONCAT('select t.name as table_name, t.timestamp as timestamp, r.id as record_id, r.timestamp as entry_date, ', IF(@sql LIKE '%id_users%', @sql, CONCAT(@sql,', -1 AS id_users')), 
+                ' from uploadTables t
+					inner join uploadRows r on (t.id = r.id_uploadTables)
+					inner join uploadCells cell on (cell.id_uploadRows = r.id)
+					inner join uploadCols col on (col.id = cell.id_uploadCols)
+					where t.id = ', table_id_param,
+					' group by t.name, t.timestamp, r.id HAVING 1 ', filter_param);
+			IF LOCATE('id_users', @sql) THEN
+				-- get user_name if there is id_users column
+				SET @sql = CONCAT('select v.*, u.name as user_name from (', @sql, ')  as v left join users u on (v.id_users = u.id)');
+			END IF;
+
+            PREPARE stmt FROM @sql;
+            EXECUTE stmt;
+            DEALLOCATE PREPARE stmt;
+        END;
+    END IF;
+END
+//
+
+DELIMITER ;
+